WIP: Blake2: Do not apply an initial key block if the key is empty
TODO: Requires adjustments of the key_block field as well. The key_block should be an `Option`

Fixes !482 with variant 2.

In theory, this is a breaking change even though usage with an empty key is kinda questionable.

It does behave consistent with other implementations. For example, Python's `hashlib` also omits the initial key block if the key is empty.